Add files via upload
This project was developed to find the optimal size of a Gamma Ray detector based on given data point. Above code set up was done in Code Block. Initially number of Gamma Ray strike was simulated and the coordinated were noted. Points.txt consists of cordinates (x,y) and the strike number. 
This algorithms tries to find optimal Radius of a gamma ray detector that can report maximum strike.